requirements.txt
Django~=5.1.1
django-mptt~=0.16.0